TransLingua/requirements.txt
streamlit
google-generativeai
python-dotenv
google.genai